+++ src/controllers/GridController.php
@@ -1,0 +1,79 @@
+<?php
+/**
+ * Grid plugin for Craft CMS 3.x
+ *
+ * Content manage CSS grids for matrix and relation fields.
+ *
+ * @link      http://wbrowar.com
+ * @copyright Copyright (c) 2019 Will Browar
+ */
+
+namespace wbrowar\grid\controllers;
+
+use wbrowar\grid\Grid;
+
+use Craft;
+use craft\web\Controller;
+
+/**
+ * Grid Controller
+ *
+ * Generally speaking, controllers are the middlemen between the front end of
+ * the CP/website and your plugin’s services. They contain action methods which
+ * handle individual tasks.
+ *
+ * A common pattern used throughout Craft involves a controller action gathering
+ * post data, saving it on a model, passing the model off to a service, and then
+ * responding to the request appropriately depending on the service method’s response.
+ *
+ * Action methods begin with the prefix “action”, followed by a description of what
+ * the method does (for example, actionSaveIngredient()).
+ *
+ * https://craftcms.com/docs/plugins/controllers
+ *
+ * @author    Will Browar
+ * @package   Grid
+ * @since     1.0.0
+ */
+class GridController extends Controller
+{
+
+    // Protected Properties
+    // =========================================================================
+
+    /**
+     * @var    bool|array Allows anonymous access to this controller's actions.
+     *         The actions must be in 'kebab-case'
+     * @access protected
+     */
+    protected $allowAnonymous = ['index', 'do-something'];
+
+    // Public Methods
+    // =========================================================================
+
+    /**
+     * Handle a request going to our plugin's index action URL,
+     * e.g.: actions/grid/grid
+     *
+     * @return mixed
+     */
+    public function actionIndex()
+    {
+        $result = 'Welcome to the GridController actionIndex() method';
+
+        return $result;
+    }
+
+    /**
+     * Handle a request going to our plugin's actionDoSomething URL,
+     * e.g.: actions/grid/grid/do-something
+     *
+     * @return mixed
+     */
+    public function actionDoSomething()
+    {
+        $result = 'Welcome to the GridController actionDoSomething() method';
+
+        return $result;
+    }
+}