+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 1.0.0 - 2019-02-02
+## 1.0.0-beta.1 - 2019-02-02
+### Added
+- Grid fields can now be used in Matrix fields to create multiple layouts for a target field
+- Added mobile layout for settings and field input
+- Grid now resaves all elements when a layout breakpoint changes in field settings
+
+## 1.0.0-beta.0 - 2019-02-02
 ### Added
 - Initial release